default
commit

+++ src/main/java/com/example/house/entity/resp/RuleResp.java
@@ -52,8 +52,8 @@ public class RuleResp implements Serializable{
     private String creator;
 
     @ApiModelProperty("版本号")
-    private long version;
+    private Integer version;
 
     @ApiModelProperty("逻辑删除状态 0未删除 1已删除")
-    private long isDelete;
+    private Integer isDelete;
 }
